backend/requirements.txt
# FastAPI & ASGI
fastapi==0.115.6
uvicorn[standard]==0.34.0
gunicorn==23.0.0

# MongoDB
motor==3.7.0
pymongo==4.10.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Validation & Settings
pydantic[email]==2.10.4
pydantic-settings==2.7.1
email-validator==2.2.0

# Google Gemini AI
google-genai==1.7.0

# HTTP Client (for LLM calls)
httpx==0.28.1

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
httpx==0.28.1

# Utilities
python-multipart==0.0.20
python-dotenv==1.0.1
